+++ examples/custom_instance_settings/outputs.tf
@@ -27,6 +27,7 @@ output "this_rds_cluster_database_name" {
 output "this_rds_cluster_master_password" {
   description = "The master password"
   value       = module.aurora.this_rds_cluster_master_password
+  sensitive   = true
 }
 
 output "this_rds_cluster_port" {
@@ -45,13 +46,9 @@ output "this_rds_cluster_instance_endpoints" {
   value       = module.aurora.this_rds_cluster_instance_endpoints
 }
 
-output "this_rds_cluster_instance_ids" {
-  description = "A list of all cluster instance ids"
-  value       = module.aurora.this_rds_cluster_instance_ids
-}
-
 // aws_security_group
 output "this_security_group_id" {
   description = "The security group ID of the cluster"
   value       = module.aurora.this_security_group_id
 }
+